force generate in ci
