+++ .github/workflows/lint.yaml
@@ -17,10 +17,17 @@ jobs:
         with:
           go-version: 1.16.x
 
+      - name: Install Protoc
+        uses: arduino/setup-protoc@v1
+        with:
+          version: '3.x'
+
       - name: Checkout code
         uses: actions/checkout@v2
 
-      - run: make generate
+      - run: make install
+      - run: make generate -B
+
       - run: git diff --exit-code
 
   lint:
