+++ .gitignore
@@ -12,3 +12,4 @@ torrents
 big_file.bin
 .leveldb/
 testdata/big_file.*
+*.bolt
