+++ .mk/dev.mk
@@ -1,0 +1,4 @@
+install:
+	go install google.golang.org/protobuf/cmd/protoc-gen-go
+
+.PHONY:: install
